[Refactor] 유저 정보 뷰 & 성향 태그 뷰 분리 (#5)
