+++ DATEROAD-SwiftUI/DATEROAD-SwiftUI/Presentation/Profile/View/EditProfileView.swift
@@ -1,0 +1,18 @@
+//
+//  EditProfileView.swift
+//  DATEROAD-SwiftUI
+//
+//  Created by 윤희슬 on 11/5/24.
+//
+
+import SwiftUI
+
+struct EditProfileView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    EditProfileView()
+}
